cmdbuf: obsolet memset to be deleted  in psb_cmdbuf.c lnc_cmdbuf.c pnw_cmdbuf.c
BZ: 9354

In psb_video code, there are obsolete memset functions called when encodingone slice.
one obsolete memset that is to init struct that is already initializedby calloc funtion.
Another obsolete memeset is to remove the reason is obsoletefields in drm_psb_cmdbuf_arg,
so to remove the obsolet the fields.

Change-Id: I7be2965d08b1af48a793620a7b469d4ae444daff
Signed-off-by: xzhai1 <xiaogang.zhai@intel.com>
Reviewed-on: http://android.intel.com:8080/21957
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>