JPEG Encode Drivers' Quantization Table dumping code needs to be improved for logcat.
BZ: 92192

Issue: Since logcat always adds a '\n' to the
end of every item of message, the current dumping
code outputs Quantization Table, which are two 8*8
tables, in an ugly format in the case that they are
dumped to logcat.

Solution: So both CTP's and MERR's JPEG Encode
Drivers' Quantization Table dumping code needs to
be improved to make the dumped info easier to read
if to logcat.

Change-Id: I81709242b7fda3fd7bf9ce898683b218ca267d06
Signed-off-by: SUN,Jing <jing.a.sun@intel.com>
Reviewed-on: http://android.intel.com:8080/96079
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>