Encode with IED enabled:increase the size of coded slice to including the padding bytes
BZ: 123356

Tangier IED encryption operates on the chunks with 16bytes, and we must include
the extra bytes beyond slice data as a whole chunk for decryption

Per IMG ticket 26132, topaz always outputs 16 byte aligned data. It pads it up to
16 bytes by inserting stuffing bytes depending on the standard being used.
This stuffing byte is 0xFF for MPEG4 and Ox00 for all other standards.

Need to handle it according to IED status (enable or disable), but from UC level,
IED is only enabled for DRM over WiDi and only uses H264 BP. So here simply includes
the padding for H264 BP always, regardless IED enabled or disabled

Change-Id: Ic5b7e03c1874aafbb73c560754485796a95ef712
Signed-off-by: Austin Yuan <shengquan.yuan@intel.com>
Reviewed-on: http://android.intel.com:8080/118880
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>